requirements.txt
# ===================================
# Dependências Core
# ===================================
python-dotenv>=1.0.0

# ===================================
# LangChain
# ===================================
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1

# ===================================
# OpenAI
# ===================================
openai>=1.0.0

# ===================================
# Vector Store
# ===================================
chromadb>=0.4.22

# ===================================
# Processamento de PDF
# ===================================
pypdf>=3.17.0

# ===================================
# SQLite (fallback para Streamlit Cloud)
# ===================================
pysqlite3-binary>=0.5.0

# ===================================
# Frontend
# ===================================
streamlit>=1.29.0

# ===================================
# Opcional - Backend API (FastAPI)
# ===================================
# Descomente as linhas abaixo se for usar a versão FastAPI
# fastapi>=0.109.0
# uvicorn[standard]>=0.27.0
# python-multipart>=0.0.6
# requests>=2.31.0

# ===================================
# Desenvolvimento e Testes
# ===================================
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0  # Linter e formatter
